requirements.txt
arrow==0.15.2
beautifulsoup4==4.8.0
bs4==0.0.1
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
decorator==4.4.0
Flask==1.1.1
Flask-cors==3.0.8
future==0.17.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
jsonpickle==1.4.1
lxml==4.4.1
MarkupSafe==1.1.1
oauth2client==4.1.3
Pillow==6.2.0
proglog==0.1.9
pyasn1==0.4.7
pyasn1-modules==0.2.6
python-dateutil==2.8.0
pytube3==9.6.4
redis==3.3.8
requests==2.22.0
rq==1.1.0
rq-dashboard==0.5.3
rsa==4.0
six==1.12.0
soupsieve==1.9.3
tqdm==4.35.0
uritemplate==3.0.0
urllib3==1.25.3
uwsgi==2.0.18
Werkzeug==0.15.6
wheel==0.33.6